Fix: type-safe strict comparisons and exit after redirect (#77)
Address Copilot review feedback on PR #80:
- Cast $_POST['post_ID'] with absint/wp_unslash before strict comparison
- Cast get_wpcm_stats_value() result to int for sub ID comparison
- Use wp_get_theme()->get('Name') instead of comparing object to string
- Cast both sides of post ID selected comparisons to int
- Cast selected/key to string in wpcm_form_dropdown for type safety
- Add exit after wp_safe_redirect() to halt execution

+++ includes/admin/post-types/meta-boxes/class-wpcm-meta-box-match-players.php
@@ -174,7 +174,7 @@ class WPCM_Meta_Box_Match_Players {
 				<option value="-1"><?php esc_html_e( 'None', 'wp-club-manager' ); ?></option>
 
 				<?php foreach ( $subs as $sub ) { ?>
-				<option value="<?php echo esc_attr( $sub->ID ); ?>"<?php echo ( get_wpcm_stats_value( $players['subs'], $id, 'sub' ) === $sub->ID ? ' selected' : '' ); ?>>
+				<option value="<?php echo esc_attr( $sub->ID ); ?>"<?php echo ( (int) get_wpcm_stats_value( $players['subs'], $id, 'sub' ) === (int) $sub->ID ? ' selected' : '' ); ?>>
 					<?php echo esc_html( get_post_meta( $sub->ID, 'wpcm_number', true ) ); ?>. <?php echo esc_html( $sub->post_title ); ?>
 				</option>
 				<?php } ?>
